Recurse on alternate recipes as well
This is so we build a full production tree for all possible recipe
choices at each stage so the UI can freely swap between alternates
without needing to rebuild part of the production tree.

The only awkward part of this is that the summary gets a little more
complicated. If we summarize from every path, then we're going to
report needing significantly more raw materials than are actually
needed (because it will be the sum of raw materials needed for every
possible path).

This probably means that generating the summary needs to be done for a
particular production tree, which probably means it needs to be done
separately from building the tree in the first place.