Implement context aware control updating
The controls are interrelated and to some degree calculated using
other data that we don't have in this specific context. Rather than
trying to somehow get that data, I think it makes more sense to set
the value that needs to be calculated to `nil` and then allow that
calculation to happen elsewhere. This way, the presence of a value
indicates which of the ratio or specific controls is the current
anchor that should be used to calculate the other.

Also, changing the timescale should change the specific control if
it's the current anchor, but if the ratio is the current anchor
there's nothing to change.