Rework production-tree to accept a recipe-id instead of an item-id
Since we are putting recipe-ids in the recipe picker dialog, it is
much more convenient to calculate a production tree based on a recipe
id. It's slightly more awkward because it requires having both a
recipe map indexed by recipe id and a grouped recipe map indexed by
result id tho.